webclient/requirements.txt
celery==3.1.22
flask
redis
pymongo==3.2.2
flask_restful
bcrypt
PyJWT